fix : 기존 pages폴더안에 있던 type 파일 type폴더안으로 이동
