--- pages/search/[nickname]/summoner/props.ts
@@ -1,7 +1,0 @@
-export const AsyncBoundaryPropsMapper = (nickname: string) => ({
-  key: nickname,
-});
-
-export const SectionPropsMapper = (nickname: string) => ({
-  nickname,
-});
